fix: make props reactive
According to https://www.solidjs.com/tutorial/props_defaults, "it is
very important to not just destructure props objects, as that would lose
reactivity if not done within a tracking scope". For this reason, when
the component is mounting, we use Object.defineProperty to create
kebab-case-formatted attributes with the descriptor of their associated
camelCase-formatted attribute.